requirements.txt
streamlit
cohere
python-dotenv
PyPDF2
Pillow
transformers
torch